requirements.txt
Django==1.9.4
djangorestframework==3.3.3
Pillow==3.2.0
psycopg2==2.6.1
requests==2.9.1
